projects/PJ2/s03p22a302/backend/requirements.txt
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.6.0
Django==2.2.7
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
idna==2.10
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
